Metadata Extractor/requirements.txt
ExifRead==3.0.0
mutagen==1.47.0
pandas==2.2.2
Pillow==10.4.0
PyPDF2==3.0.1
streamlit==1.29.0
